requirements.txt
# Cross-platform hotkey support
pynput>=1.7.6      # For macOS and Linux
keyboard>=0.13.5    # For Windows

# Audio processing
faster-whisper==0.10.0
kittentts==0.1.0
# pyaudio==0.2.11  # Commented out - use pipwin on Windows, pip on macOS/Linux
soundfile==0.12.1

# Data processing
openpyxl==3.1.2
requests==2.31.0
python-dateutil==2.8.2

# AI/LLM
openai>=1.0.0
python-dotenv==1.0.0

# Additional dependencies
numpy>=1.21.0

